MAINT: split out misc functions into utils and widgetgroups into their own file
